Reduce feature gates amount
